+++ server/src/controller/user.js
@@ -28,15 +28,18 @@ const verifyUser = async(req, res) => {
     //find if the user exists
    const data = await User.findOne({$or:[{email:  req.body.email}, {userName: req.body.userName }, {phoneNumber:req.body.phoneNumber}]})
    if(data){
+       //db password ---->compare------> 
         const isMatched =await bcrypt.compare(req.body.password, data.password); // false
         //generate a jwt token for him
         const {password, ...allOtherItem} = req.body
-        const token = await jwt.sign(allOtherItem, process.env.SECRET_KEY, { expiresIn: '1h'  });
+        const token = await jwt.sign(allOtherItem, process.env.SECRET_KEY, { expiresIn: '12h'  });
         if(isMatched && token){
             res.json({
                 msg: "login success",
                 isLoggedIn: true,
-                token: token
+                token: token,
+                id: data._id,
+                role: data.role
             })
         }else{
             res.json({
