+++ server/src/model/user.js
@@ -3,6 +3,7 @@ const UserSchema = new mongoose.Schema({
   firstName: String,
   lastName: String,
   password: {type: String, required: true},
+  role: {type: String, default: 'user'},
   email:  {type: String, default:''},
   phoneNumber: {type: String, default:''},
   userName:  {type: String, default:''},
